pyproject.toml
[project]
name = "awesome-ai-agents"
version = "0.1.0"
description = "A curated collection of open-source AI agents built with LangGraph v1 for learning and experimentation."
readme = "README.md"
authors = [
    { name = "Lokeswaran Aruljothy", email = "lokeswaranaruljothy@gmail.com" },
]
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.3",
    "langchain-anthropic>=1.0.1",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.7",
]
